refactor: rearrange wallet information components
